requirements.txt
Flask
sumy
python-docx
werkzeug